+++ klause/src/commonMain/kotlin/com/eignex/klause/backtrack/selector/IndomainMiddle.kt
@@ -10,8 +10,8 @@ import kotlin.random.Random
  * from [IndomainMedian] (middle *by position*) on skewed or holey domains.
  */
 object IndomainMiddle : ValueSelector {
-    override fun values(session: PropagationSession, varRef: VarRef, rng: Random): Sequence<Int> = when (varRef) {
-        is VarRef.Bool -> sequenceOf(0, 1)
+    override fun values(session: PropagationSession, varRef: VarRef, rng: Random): Sequence<Long> = when (varRef) {
+        is VarRef.Bool -> sequenceOf(0L, 1L)
 
         is VarRef.IntVar -> {
             val d = session.intDomain(varRef.varId)
